# app/services/simulator-api/docs/swagger.yaml
basePath: /
definitions:
  responses.Failed:
    properties:
      error:
        type: string
      status:
        example: ERROR
        type: string
    type: object
  responses.Health:
    properties:
      buildVersion:
        example: localhost
        type: string
    type: object
  responses.Success:
    properties:
      status:
        example: OK
        type: string
    type: object
  responses.SuccessHealth:
    properties:
      health:
        $ref: '#/definitions/responses.Health'
      status:
        example: OK
        type: string
    type: object
  responses.SuccessLogin:
    properties:
      status:
        example: OK
        type: string
      token:
        type: string
    type: object
  responses.Validator:
    properties:
      details:
        items:
          type: string
        type: array
      status:
        example: ERROR
        type: string
    type: object
  webmodels.LoginUser:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  webmodels.RegisterUser:
    properties:
      company:
        type: string
      email:
        type: string
      firstName:
        type: string
      language:
        enum:
        - es
        - en
        - fr
        - pt
        maxLength: 2
        type: string
      lastName:
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    required:
    - company
    - email
    - firstName
    - language
    - lastName
    - password
    type: object
info:
  contact: {}
  description: Devices Simulator documentation API.
  title: Swagger MYC-DEVICE-SIMULATOR API
  version: "1.0"
paths:
  /api/health:
    get:
      consumes:
      - application/json
      description: Get status service to be alive
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessHealth'
      summary: Health check EndPoint
      tags:
      - Health
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login authentication user in platform
      parameters:
      - description: LoginUser
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/webmodels.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessLogin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Validator'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Failed'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Failed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Failed'
      summary: Login user EndPoint
      tags:
      - Auth
  /api/v1/users:
    post:
      consumes:
      - application/json
      description: Create a new user in the system.
      parameters:
      - description: UserRegister
        in: body
        name: UserRegister
        required: true
        schema:
          $ref: '#/definitions/webmodels.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Validator'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/responses.Failed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Failed'
      summary: Create user registration EndPoint
      tags:
      - Users
  /api/v1/users/activate/{activateToken}:
    post:
      description: Activation a user in the system.
      parameters:
      - description: ActivateToken
        in: path
        name: activateToken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Validator'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Failed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Failed'
      summary: Activate user activation EndPoint
      tags:
      - Users
swagger: "2.0"